// frontend/src/services/models/entry-create.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Plant Tracker API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface EntryCreate
 */
export interface EntryCreate {
    /**
     * 
     * @type {string}
     * @memberof EntryCreate
     */
    'timestamp': string;
    /**
     * 
     * @type {Array<string>}
     * @memberof EntryCreate
     */
    'activities': Array<string>;
    /**
     * 
     * @type {string}
     * @memberof EntryCreate
     */
    'plant_id': string;
    /**
     * 
     * @type {string}
     * @memberof EntryCreate
     */
    'notes': string | null;
    /**
     * 
     * @type {number}
     * @memberof EntryCreate
     */
    'plant_health': number;
}